+++ antarest/storage/repository/filesystem/raw_file_node.py
@@ -44,5 +44,5 @@ class RawFileNode(INode[str, str, str]):
                 f"url should be fully resolved when arrives on {self.__class__.__name__}"
             )
 
-    def unloaded_placeholder_name(self):
+    def _unloaded_placeholder_name(self) -> str:
         return "RawFileNode"
