+++ webapp/src/components/StudyView/StudyDataView/StudyFileView.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable react-hooks/exhaustive-deps */
 import React, { useEffect, useState } from 'react';
 import { Theme, createStyles, makeStyles } from '@material-ui/core';
 import { useSnackbar } from 'notistack';
